build: lint every platform the reference claims
A file behind a build tag is invisible to a lint run under another GOOS, so a
pass on one platform said nothing about the others and the pipeline was the
first thing to read them. `task lint` runs once per operating system named in
docs/reference/platforms.md, which is what makes a local pass mean what a
pipeline pass means.

Two defects it found on the first run.

internal/fsx/class_linux.go converted the statfs type before switching on it.
Every architecture the platform reference claims is 64-bit, where the field is
already int64, so the conversion described a portability the support matrix does
not claim.

buildAgentfs lived in pipe_test.go, which is built only where signals are, while
docs_test.go calls it unconditionally — so the root test package did not compile
under Windows. Windows is compile-only in the matrix, so no test run reported it.
The helper moves to binary_test.go with no build tag, because a helper reached
from an untagged file has to be as portable as its callers, and it names the
executable the way the host does.

+++ internal/fsx/class_linux.go
@@ -25,7 +25,7 @@ func classify(dir string) Filesystem {
 	if err := syscall.Statfs(dir, &st); err != nil {
 		return Filesystem{Kind: KindUnknown, Type: "unprobeable"}
 	}
-	switch int64(st.Type) {
+	switch st.Type {
 	case magicEXT:
 		return Filesystem{Kind: KindLocal, Type: "ext"}
 	case magicXFS:
